Specify scalar size as float32

+++ babyvec/models.py
@@ -4,7 +4,7 @@ import numpy as np
 import numpy.typing as npt
 
 
-EmbeddingScalarType = np.float64
+EmbeddingScalarType = np.float32
 Embedding = npt.NDArray[EmbeddingScalarType]
 EmbeddingId = int
